First commit since stupid bug with adding/removing entities from implicit grid. It works now, but code is still incredibly messy, definitely more so than before the problem arose. Unsure of how I'll use Entity's/DynamicEntity's instance variables "offScreen" and "removed", from here on out, since the problem seemed to stem from the fact that the old way of just setting removed to true, and letting level handle the lists sorting doesn't work with this new setup. That said, I can still see a usefulness of keeping those around for quick checks of whether or not the object is inside the grid, or still exists inside the level entities list, etc. etc.

+++ src/com/bradboughn/rain/camera/Camera.java
@@ -93,8 +93,8 @@ public class Camera
 //                if (offsetY > player.getY() - Camera.height/2 && ytick > 30) offsetY--;
 //            } else ytick = 0;
 //        }
-            offsetX = player.getX() - (Camera.width>>1);
-            offsetY = player.getY() - (Camera.height>>1);
+            offsetX = (int)player.getX() - (Camera.width>>1);
+            offsetY = (int)player.getY() - (Camera.height>>1);
 
          
             
@@ -103,7 +103,9 @@ public class Camera
     
     public static void update()
     {
+
         updateStartEndColRow();
+        
     }
    
     public static void render()
