+++ src/com/bradboughn/rain/entity/spawner/ParticleSpawner.java
@@ -9,7 +9,7 @@ public class ParticleSpawner extends Spawner
     
     private int life;
     
-    public ParticleSpawner(int x, int y, Type type, int life, int amount, Level level) 
+    public ParticleSpawner(double x, double y, Type type, int life, int amount, Level level) 
     {
         super(x, y, Type.PARTICLE_WIZ, amount, level);
         this.life = life;
